Add files via upload
In this code, the concept of hashing is implemented, with the following methods :-
get() - to get the value of the provided key
put() - to add key to the hash table
remove() - to remove the key from the hash table
containsKey() - to check whether the hash table contains the provided key or not
hashFunc() - to calculate the index of the key to be stored in the hash table
isEmpty() - to check whether the hash table is empty or not
rehash() - to rehash the hash table, i.e. double the size of the hash table when the load factor value increases by the threshold value. 
keySet() - to return an ArrayList of all the keys in the hash table